Restructure source files & setup public directory
