+++ public/stylesheets/style.css
@@ -1,0 +1,3 @@
+body {
+  text-align: center;
+}
